backend/voice-assistant/requirements.txt
python-dotenv==1.0.0
livekit-agents[images]~=1.5
livekit-plugins-silero~=1.5
livekit-plugins-google~=1.5
livekit-plugins-turn-detector~=1.5
livekit-plugins-dtln>=0.1
google-generativeai==0.8.3
google-cloud-aiplatform>=1.65.0
flask==3.0.0
gunicorn==21.2.0
